refactored the language detection and translation into one class
